employee_project/requirements.txt
Django>=4.0,<5.0
djangorestframework
drf-yasg
django-environ
psycopg2-binary
Faker
django-filter
